+++ py/send_message.py
@@ -1014,7 +1014,7 @@ class SendMessage:
         # pylint: disable=line-too-long
 
         inp = input(
-            "Select one of: 1=normal; 2=erc20; 3=erc721; 4=unknown erc20; 5=large data field; 6=BSC; 7=unknown network; 8=eip1559: "
+            "Select one of: 1=normal; 2=erc20; 3=erc721; 4=unknown erc20; 5=large data field; 6=BSC; 7=unknown network; 8=eip1559; 9=Arbitrum: "
         ).strip()
 
         chain_id = 1  # mainnet
@@ -1022,8 +1022,10 @@ class SendMessage:
             chain_id = 56
         elif inp == "7":
             chain_id = 123456
+        elif inp == "9":
+            chain_id = 42161  # Arbitrum One
 
-        if inp in ("1", "6", "7"):
+        if inp in ("1", "6", "7", "9"):
             # fmt: off
             tx = bytes([0xf8, 0x6e, 0x82, 0x1f, 0xdc, 0x85, 0x01, 0x65, 0xa0, 0xbc, 0x00, 0x82, 0x52,
             0x08, 0x94, 0x04, 0xf2, 0x64, 0xcf, 0x34, 0x44, 0x03, 0x13, 0xb4, 0xa0, 0x19, 0x2a,