+++ CHANGELOG.md
@@ -7,6 +7,7 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 ## Firmware
 
 ### [Unreleased]
+- Ethereum: add confirmation screen for known L2 networks and sidechains and change their unit to ETH
 
 ### 9.22.0
 - Update manufacturer HID descriptor to bitbox.swiss